fixed rich-text view
