+++ front-end/package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^4.3.1",
     "date-fns": "^2.8.1",
     "draft-convert": "^2.1.8",
+    "draft-js": "^0.11.5",
     "medium-draft": "^0.5.18",
     "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.16",
