+++ front-end/src/Views/BlogPage/components/BlogPost/BlogPost.jsx
@@ -1,14 +1,22 @@
 import React from "react";
 import { Card, Row, Badge, Col } from "react-bootstrap";
 import { IndexLinkContainer } from "react-router-bootstrap";
+import {convertToHTML} from "draft-convert";
+
+const stripFromTags = str => {
+  console.log(str);
+  str = str.replace(/(<([^>]+)>)|(<([^>]+)\.\.\.)/gi, '');
+  return str;
+};
 
 const BlogPost = ({data, ...rest}) => {
+  const str = stripFromTags(data.summary);
   return (
     <IndexLinkContainer to={`/blog/blogpost/${data.id}`}>
       <Row {...rest}>
         <Col>
           <h3>{data.category}</h3>
-          <p>{`${data.summary.substring(0, 100)}...`}</p>
+          <p>{`${str.substring(0, 100)}...`}</p>
           <Row>
             <Col>
               <small>Tagi:</small>
